+++ src/Symfony/Bundle/FrameworkBundle/CHANGELOG.md
@@ -30,6 +30,8 @@ CHANGELOG
  * Register `JsonStreamerBundle` automatically when the JsonStreamer component is installed
  * Make `framework.asset_mapper` an alias of the `asset_mapper` configuration provided by `AssetMapperBundle`
  * Register `AssetMapperBundle` automatically when the AssetMapper component is installed
+ * Make `framework.rate_limiter` an alias of the `rate_limiter` configuration provided by `RateLimiterBundle`
+ * Register `RateLimiterBundle` automatically when the RateLimiter component is installed
  * Make `framework.workflows` an alias of the `workflow` configuration provided by `WorkflowBundle`
  * Register `WorkflowBundle` automatically when the Workflow component is installed
  * Make `framework.remote_event` an alias of the `remote_event` configuration provided by `RemoteEventBundle`